test(errors): split tests into classification + exit_code modules
tools/errors/tests.rs (478) split into classification_tests (classify kind +
next_step hints, 268) and exit_code_tests (ErrorCode->exit code contract +
retryable, 211). T-ER IDs and fn names are preserved since docs/decisions and
docs/audit reference them as exit-code contract pins. Production errors.rs is
unchanged apart from the mod declarations; 449 tests pass.
